backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq
langchain-google-community
langchain-text-splitters
langgraph
langgraph-prebuilt
faiss-cpu
google-generativeai
google-api-python-client
groq
pymupdf
beautifulsoup4
requests
gTTS
python-dotenv
